Deduplicate Dav1dTileState and nested structs
