+++ src/thread_task.rs
@@ -110,29 +110,6 @@ use crate::src::levels::BlockSize;
 
 use crate::src::internal::Dav1dTaskContext;
 
-#[derive(Copy, Clone)]
-#[repr(C)]
-pub struct Dav1dTileState {
-    pub cdf: CdfContext,
-    pub msac: MsacContext,
-    pub tiling: Dav1dTileState_tiling,
-    pub progress: [atomic_int; 2],
-    pub frame_thread: [Dav1dTileState_frame_thread; 2],
-    pub lowest_pixel: *mut [[libc::c_int; 2]; 7],
-    pub dqmem: [[[uint16_t; 2]; 3]; 8],
-    pub dq: *const [[uint16_t; 2]; 3],
-    pub last_qidx: libc::c_int,
-    pub last_delta_lf: [int8_t; 4],
-    pub lflvlmem: [[[[uint8_t; 2]; 8]; 4]; 8],
-    pub lflvl: *const [[[uint8_t; 2]; 8]; 4],
-    pub lr_ref: [*mut Av1RestorationUnit; 3],
-}
-#[derive(Copy, Clone)]
-#[repr(C)]
-pub struct Dav1dTileState_frame_thread {
-    pub pal_idx: *mut uint8_t,
-    pub cf: *mut libc::c_void,
-}
 use crate::src::internal::Dav1dTileState_tiling;
 use crate::src::internal::Dav1dContext;
 use crate::src::intra_edge::EdgeFlags;
@@ -519,6 +496,8 @@ unsafe extern "C" fn dav1d_set_thread_name(name: *const libc::c_char) {
 use crate::include::common::intops::iclip;
 use crate::include::common::intops::imax;
 use crate::include::common::intops::umin;
+
+use super::internal::Dav1dTileState;
 #[inline]
 unsafe extern "C" fn reset_task_cur(
     c: *const Dav1dContext,
